raw_errors_bitmask issue solved and tested

+++ components/jk_rs485_bms/jk_rs485_bms.h
@@ -726,7 +726,7 @@ class JkRS485Bms : public PollingComponent, public jk_rs485_sniffer::JkRS485Snif
   void decode_jk04_cell_info_(const std::vector<uint8_t> &data);
   void decode_device_info_(const std::vector<uint8_t> &data);
 
-  std::string error_bits_to_string_(uint16_t bitmask);
+  std::string error_bits_to_string_(uint32_t bitmask);
   std::string mode_bits_to_string_(uint16_t bitmask);
 
   float get_temperature_(const uint16_t value) {
